Create utils/api.ts for fetchApi and related functions
This was originally created in support-api.ts to try out using react-query. Now that we're moving towards that as the standard approach, it should have its own file.
